pyproject.toml
[project]
name = "slide-generation-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.73.0",
    "dotenv>=0.9.9",
    "python-pptx>=1.0.2",
    "streamlit>=1.51.0",
    "PyMuPDF>=1.23.0",
]
